requirements.txt
django
channels
daphne
djangorestframework
channels-redis
psycopg2-binary
dj-database-url